[Maintenance mode] Merged the maintenance mode activation/deactivation in one command
