--- bundles/CoreBundle/Command/MaintenanceModeEnableCommand.php
@@ -1,44 +1,0 @@
-<?php
-/**
- * Pimcore
- *
- * This source file is available under two different licenses:
- * - GNU General Public License version 3 (GPLv3)
- * - Pimcore Enterprise License (PEL)
- * Full copyright and license information is available in
- * LICENSE.md which is distributed with this source code.
- *
- * @copyright  Copyright (c) Pimcore GmbH (http://www.pimcore.org)
- * @license    http://www.pimcore.org/license     GPLv3 and PEL
- */
-
-namespace Pimcore\Bundle\CoreBundle\Command;
-
-use Exception;
-use Pimcore\Console\AbstractCommand;
-use Pimcore\Tool\Admin;
-use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Output\OutputInterface;
-
-class MaintenanceModeEnableCommand extends AbstractCommand
-{
-    /**
-     * @inheritDoc
-     */
-    protected function configure()
-    {
-        $this
-            ->setName('pimcore:maintenance-mode:enable')
-            ->setDescription('Enable maintenance mode');
-    }
-
-    /**
-     * @inheritDoc
-     * @throws Exception
-     */
-    protected function execute(InputInterface $input, OutputInterface $output)
-    {
-        Admin::activateMaintenanceMode('command-line-dummy-session-id');
-        return 0;
-    }
-}
